refactor(rpc): split wire types and tests
