+++ bedrock/src/transactions/rpc.rs
@@ -13,7 +13,7 @@ use crate::{
     },
     smart_account::{SafeSmartAccountError, UserOperation},
 };
-use alloy::primitives::{Address, Bytes, FixedBytes, B256, U128, U256};
+use alloy::primitives::{Address, Bytes, FixedBytes, B256};
 use alloy::sol_types::SolCall;
 use alloy::{hex::FromHex, providers::MULTICALL3_ADDRESS};
 
@@ -23,105 +23,20 @@ use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use std::sync::{Arc, OnceLock};
 
+mod wire;
+
+pub(crate) use wire::{ErrorPayload, JsonRpcRequest};
+pub use wire::{
+    Id, PmSponsorUserOperationResponse, RelaySafeTransactionRequest, RpcMethod,
+    RpcProviderName, SponsorUserOperationResponse, SponsorshipContext,
+    WaGetUserOperationReceiptResponse,
+};
+
+#[cfg(test)]
+mod tests;
+
 /// Global RPC client instance for Bedrock operations
 static RPC_CLIENT_INSTANCE: OnceLock<RpcClient> = OnceLock::new();
-
-/// JSON-RPC request ID
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum Id {
-    /// Numeric ID
-    Number(u64),
-    /// String ID
-    String(String),
-}
-
-/// Supported RPC methods in Bedrock
-#[derive(Debug, Clone, Serialize)]
-pub enum RpcMethod {
-    /// Request sponsorship for a `UserOperation` (V1)
-    #[serde(rename = "wa_sponsorUserOperation")]
-    SponsorUserOperation,
-    /// Request sponsorship for a `UserOperation` (V2)
-    #[serde(rename = "pm_sponsorUserOperation")]
-    PmSponsorUserOperation,
-    /// Queries the status of a `UserOperation`
-    #[serde(rename = "wa_getUserOperationReceipt")]
-    WaGetUserOperationReceipt,
-    /// Relay a signed Safe `execTransaction` for the backend to submit on-chain
-    #[serde(rename = "wa_relaySafeTransaction")]
-    RelaySafeTransaction,
-    /// Submit a signed `UserOperation` (V1)
-    #[serde(rename = "eth_sendUserOperation")]
-    SendUserOperation,
-    /// Submit a signed `UserOperation` (V2)
-    #[serde(rename = "eth_sendUserOperation")]
-    SendUserOperationV2,
-    /// Make a read call to a smart contract
-    #[serde(rename = "eth_call")]
-    EthCall,
-    /// Read a single storage slot of a contract
-    #[serde(rename = "eth_getStorageAt")]
-    EthGetStorageAt,
-    /// Query supported ERC-4337 entry points
-    #[serde(rename = "eth_supportedEntryPoints")]
-    SupportedEntryPoints,
-}
-
-/// 4337 provider selection to be passed by native apps
-#[derive(Debug, Clone, Copy, PartialEq, Eq, uniffi::Enum, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum RpcProviderName {
-    /// Let TFH backend load balance between available providers
-    Any,
-    /// Use Alchemy as 4337 provider
-    Alchemy,
-    /// Use Pimlico as 4337 provider
-    Pimlico,
-}
-
-impl RpcProviderName {
-    /// Returns the wire/header value for the provider
-    #[must_use]
-    pub const fn as_str(&self) -> &'static str {
-        match self {
-            Self::Any => "any",
-            Self::Alchemy => "alchemy",
-            Self::Pimlico => "pimlico",
-        }
-    }
-}
-
-/// JSON-RPC request envelope, shared with [`super::custom_bundler`].
-#[derive(Debug, Serialize)]
-pub(crate) struct JsonRpcRequest<T> {
-    jsonrpc: &'static str,
-    id: Id,
-    method: RpcMethod,
-    params: T,
-}
-
-impl<T> JsonRpcRequest<T> {
-    pub(crate) const fn new(method: RpcMethod, id: Id, params: T) -> Self {
-        Self {
-            jsonrpc: "2.0",
-            id,
-            method,
-            params,
-        }
-    }
-}
-
-/// JSON-RPC error payload, shared with [`super::custom_bundler`].
-#[derive(Debug, Deserialize)]
-pub(crate) struct ErrorPayload {
-    pub code: i64,
-    pub message: String,
-    #[serde(default)]
-    #[allow(dead_code)]
-    pub data: Option<Value>,
-}
-
 /// Errors that can occur when interacting with RPC operations.
 #[crate::bedrock_error]
 
@@ -186,148 +101,6 @@ impl From<SafeSmartAccountError> for RpcError {
     fn from(e: SafeSmartAccountError) -> Self {
         Self::SafeSmartAccountError(e.to_string())
     }
-}
-
-/// Response from `wa_sponsorUserOperation`
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SponsorUserOperationResponse {
-    /// Paymaster address
-    pub paymaster: Option<Address>,
-    /// Paymaster data
-    pub paymaster_data: Option<Bytes>,
-    /// Pre-verification gas
-    pub pre_verification_gas: U256,
-    /// Verification gas limit
-    pub verification_gas_limit: U128,
-    /// Call gas limit
-    pub call_gas_limit: U128,
-    /// Paymaster verification gas limit
-    pub paymaster_verification_gas_limit: Option<U128>,
-    /// Paymaster post-op gas limit
-    pub paymaster_post_op_gas_limit: Option<U128>,
-    /// Max priority fee per gas
-    pub max_priority_fee_per_gas: U128,
-    /// Max fee per gas
-    pub max_fee_per_gas: U128,
-    /// provider name
-    pub provider_name: RpcProviderName,
-}
-
-/// Response from `pm_sponsorUserOperation` (V2)
-///
-/// Paymaster fields (`paymaster`, `paymaster_data`,
-/// `paymaster_verification_gas_limit`, `paymaster_post_op_gas_limit`) are
-/// absent from the bundler-sponsored response shape and present on the
-/// self-sponsored (token) response — see
-/// `bedrock/src/transactions/transaction.md`. They are modelled as
-/// `Option<T>` so both shapes deserialize.
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct PmSponsorUserOperationResponse {
-    /// Call gas limit
-    pub call_gas_limit: U128,
-    /// Verification gas limit
-    pub verification_gas_limit: U128,
-    /// Pre-verification gas
-    pub pre_verification_gas: U256,
-    /// Max fee per gas
-    pub max_fee_per_gas: U128,
-    /// Max priority fee per gas
-    pub max_priority_fee_per_gas: U128,
-    /// Paymaster address (absent on the bundler-sponsored path)
-    pub paymaster: Option<Address>,
-    /// Paymaster verification gas limit (absent on the bundler-sponsored path)
-    pub paymaster_verification_gas_limit: Option<U128>,
-    /// Paymaster post-op gas limit (absent on the bundler-sponsored path)
-    pub paymaster_post_op_gas_limit: Option<U128>,
-    /// Paymaster data (absent on the bundler-sponsored path)
-    pub paymaster_data: Option<Bytes>,
-}
-
-/// Context object passed as the third parameter of `pm_sponsorUserOperation`.
-///
-/// V2 sponsorship is a two-mode protocol: an initial protocol-sponsored
-/// attempt with empty context, followed (on a structured decline) by a
-/// self-sponsored retry that names the ERC-20 token paying for gas. See
-/// `bedrock/src/transactions/transaction.md`.
-#[derive(Debug, Clone)]
-pub enum SponsorshipContext {
-    /// Empty context — request protocol sponsorship (the wallet provider
-    /// pays gas). Distinct from the ERC-4337 notion of bundler sponsorship,
-    /// which implies a user-appointed bundler choosing to sponsor.
-    Protocol,
-    /// Self-sponsored mode with the given ERC-20 token paying for gas.
-    SelfSponsoredToken(Address),
-}
-
-impl SponsorshipContext {
-    fn to_json_value(&self) -> serde_json::Value {
-        match self {
-            Self::Protocol => serde_json::json!({}),
-            Self::SelfSponsoredToken(token) => {
-                serde_json::json!({ "token": format!("{token:?}") })
-            }
-        }
-    }
-}
-
-/// Response from `wa_getUserOperationReceipt`
-#[derive(Debug, Deserialize, uniffi::Record, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct WaGetUserOperationReceiptResponse {
-    /// User operation hash
-    pub user_op_hash: String,
-    /// Transaction hash, if the user operation has been included in a block
-    pub transaction_hash: Option<String>,
-    /// Sender address
-    pub sender: String,
-    /// Status (`pending`, `error`, `mined_success`, or `mined_revert`)
-    pub status: String,
-    /// Source (flexible field representing the transaction type or origin)
-    pub source: String,
-    /// Source ID, if available
-    pub source_id: Option<String>,
-    /// Self-sponsor token, if applicable
-    pub self_sponsor_token: Option<String>,
-    /// Self-sponsor amount, if applicable
-    pub self_sponsor_amount: Option<String>,
-    /// Block timestamp, if the user operation has been included in a block
-    pub block_timestamp: Option<String>,
-}
-
-/// Single positional param for the `wa_relaySafeTransaction` JSON-RPC call.
-///
-/// Mirrors the arguments of the Safe `execTransaction` method plus the target
-/// Safe. All numeric fields are `0x`-prefixed hex and addresses are
-/// `0x`-prefixed; `signatures` is the packed owner signature(s).
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelaySafeTransactionRequest {
-    /// The Safe whose `execTransaction` is being called.
-    pub safe_address: String,
-    /// `to` argument of `execTransaction`.
-    pub to: String,
-    /// `value` argument.
-    pub value: String,
-    /// `data` argument.
-    pub data: String,
-    /// `operation` argument (`0` = call, `1` = delegatecall).
-    pub operation: u8,
-    /// `safeTxGas` argument.
-    pub safe_tx_gas: String,
-    /// `baseGas` argument.
-    pub base_gas: String,
-    /// `gasPrice` argument.
-    pub gas_price: String,
-    /// `gasToken` argument.
-    pub gas_token: String,
-    /// `refundReceiver` argument.
-    pub refund_receiver: String,
-    /// `nonce` argument.
-    pub nonce: String,
-    /// Packed owner signature(s).
-    pub signatures: String,
 }
 
 /// RPC client for handling 4337 `UserOperation` requests
@@ -767,291 +540,4 @@ pub fn get_rpc_client() -> Result<&'static RpcClient, RpcError> {
     RPC_CLIENT_INSTANCE
         .get()
         .ok_or(RpcError::HttpClientNotInitialized)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloy::primitives::{address, bytes, U256};
-    use serde_json::json;
-
-    #[test]
-    fn test_sponsor_response_parsing() {
-        let json_response = json!({
-            "paymaster": "0x0000000000000039cd5e8aE05257CE51C473ddd1",
-            "paymasterData": "0x01000066d1a1a4",
-            "preVerificationGas": "0x350f7",
-            "verificationGasLimit": "0x501ab",
-            "callGasLimit": "0x212df",
-            "paymasterVerificationGasLimit": "0x6dae",
-            "paymasterPostOpGasLimit": "0x706e",
-            "maxPriorityFeePerGas": "0x3B9ACA00",
-            "maxFeePerGas": "0x7A5CF70D5",
-            "providerName":"pimlico",
-        });
-
-        let response: SponsorUserOperationResponse =
-            serde_json::from_value(json_response).unwrap();
-
-        assert_eq!(
-            response.paymaster,
-            Some(address!("0000000000000039cd5e8aE05257CE51C473ddd1"))
-        );
-        assert_eq!(response.call_gas_limit, U128::from(0x212df));
-    }
-
-    #[test]
-    fn test_error_payload_parsing() {
-        let error_json = json!({
-            "code": -32000,
-            "message": "execution reverted",
-            "data": null
-        });
-
-        let error_payload: ErrorPayload = serde_json::from_value(error_json).unwrap();
-
-        assert_eq!(error_payload.code, -32000);
-        assert_eq!(error_payload.message, "execution reverted");
-    }
-
-    #[test]
-    fn test_user_operation_serialization_with_null_fields() {
-        let user_op = UserOperation {
-            sender: address!("5a6b47F4131bf1feAFA56A05573314BcF44C9149"),
-            nonce: U256::from_str_radix(
-                "009aaffd82ed9852f394acca3e4b30ef51880188532374de0000000000000000",
-                16,
-            )
-            .unwrap(),
-            factory: None,
-            factory_data: None,
-            call_data: bytes!("0xe9ae5c53"),
-            call_gas_limit: U128::from(0x13_880),
-            verification_gas_limit: U128::from(0x60_B01),
-            pre_verification_gas: U256::from(0xD3E3),
-            max_fee_per_gas: U128::from(0x3B9A_CA00),
-            max_priority_fee_per_gas: U128::from(0x2B9A_CA00),
-            paymaster: None,
-            paymaster_verification_gas_limit: None,
-            paymaster_post_op_gas_limit: None,
-            paymaster_data: None,
-            signature: vec![0xff; 77].into(),
-        };
-
-        // Test that UserOperation can be serialized
-        let serialized = serde_json::to_value(&user_op).unwrap();
-
-        // Print the actual serialized output to see the format
-        println!(
-            "UserOperation serialization: {}",
-            serde_json::to_string_pretty(&serialized).unwrap()
-        );
-
-        // Verify the key field is properly serialized with camelCase naming
-        assert_eq!(serialized["callData"], "0xe9ae5c53");
-        assert_eq!(serialized["callGasLimit"], "0x13880");
-        assert_eq!(serialized["maxFeePerGas"], "0x3b9aca00");
-        assert_eq!(serialized["maxPriorityFeePerGas"], "0x2b9aca00");
-        assert_eq!(
-            serialized["nonce"],
-            "0x9aaffd82ed9852f394acca3e4b30ef51880188532374de0000000000000000" // no leading zeroes
-        );
-        assert_eq!(serialized["preVerificationGas"], "0xd3e3");
-        assert_eq!(
-            serialized["sender"],
-            "0x5a6b47f4131bf1feafa56a05573314bcf44c9149"
-        );
-        assert_eq!(
-            serialized["signature"],
-            "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
-        );
-        assert_eq!(serialized["verificationGasLimit"], "0x60b01");
-
-        // Optional fields should be null when None
-        assert!(serialized["factory"].is_null());
-        assert!(serialized["factoryData"].is_null());
-        assert!(serialized["paymaster"].is_null());
-        assert!(serialized["paymasterData"].is_null());
-        assert!(serialized["paymasterVerificationGasLimit"].is_null());
-        assert!(serialized["paymasterPostOpGasLimit"].is_null());
-    }
-
-    #[test]
-    fn test_user_operation_serialization_with() {
-        let user_op = UserOperation {
-            sender: address!("5a6b47F4131bf1feAFA56A05573314BcF44C9149"),
-            nonce: U256::from_str_radix(
-                "009aaffd82ed9852f394acca3e4b30ef51880188532374de0000000000000000",
-                16,
-            )
-            .unwrap(),
-            factory: Some(address!("0x0000000071727De22E5E9d8BAf0edAc6f37da032")),
-            factory_data: Some(bytes!("0x7702")),
-            call_data: bytes!("0xe9ae5c53"),
-            call_gas_limit: U128::from(0x13_880),
-            verification_gas_limit: U128::from(0x60_B01),
-            pre_verification_gas: U256::from(0xD3E3),
-            max_fee_per_gas: U128::from(0x3B9A_CA00),
-            max_priority_fee_per_gas: U128::from(0x2B9A_CA00),
-            paymaster: Some(address!("0xEF725Aa22d43Ea69FB22bE2EBe6ECa205a6BCf5B")),
-            paymaster_verification_gas_limit: Some(U128::from(10)),
-            paymaster_post_op_gas_limit: Some(U128::from(0)),
-            paymaster_data: Some(bytes!("0x")),
-            signature: vec![0xff; 77].into(),
-        };
-
-        // Test that UserOperation can be serialized
-        let serialized = serde_json::to_value(&user_op).unwrap();
-
-        // Print the actual serialized output to see the format
-        println!(
-            "UserOperation serialization: {}",
-            serde_json::to_string_pretty(&serialized).unwrap()
-        );
-
-        // Verify the key field is properly serialized with camelCase naming
-        assert_eq!(serialized["callData"], "0xe9ae5c53");
-        assert_eq!(serialized["callGasLimit"], "0x13880");
-        assert_eq!(serialized["maxFeePerGas"], "0x3b9aca00");
-        assert_eq!(serialized["maxPriorityFeePerGas"], "0x2b9aca00");
-        assert_eq!(
-            serialized["nonce"],
-            "0x9aaffd82ed9852f394acca3e4b30ef51880188532374de0000000000000000" // no leading zeroes
-        );
-        assert_eq!(serialized["preVerificationGas"], "0xd3e3");
-        assert_eq!(
-            serialized["sender"],
-            "0x5a6b47f4131bf1feafa56a05573314bcf44c9149"
-        );
-        assert_eq!(
-            serialized["signature"],
-            "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
-        );
-        assert_eq!(serialized["verificationGasLimit"], "0x60b01");
-
-        assert_eq!(
-            serialized["factory"],
-            "0x0000000071727De22E5E9d8BAf0edAc6f37da032"
-        );
-        assert_eq!(serialized["factoryData"], "0x7702");
-        assert_eq!(
-            serialized["paymaster"],
-            "0xEF725Aa22d43Ea69FB22bE2EBe6ECa205a6BCf5B"
-        );
-        assert_eq!(serialized["paymasterData"], "0x");
-        assert_eq!(serialized["paymasterVerificationGasLimit"], "0xa");
-        assert_eq!(serialized["paymasterPostOpGasLimit"], "0x0");
-    }
-
-    #[test]
-    fn test_rpc_endpoint_v2_methods_route_to_v2() {
-        let network = Network::WorldChain;
-        for method in [
-            RpcMethod::PmSponsorUserOperation,
-            RpcMethod::SendUserOperationV2,
-            RpcMethod::EthCall,
-        ] {
-            let url = RpcClient::rpc_endpoint(network, &method);
-            assert!(
-                url.starts_with("/v2/"),
-                "{method:?} should route to /v2/, got {url}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_rpc_endpoint_legacy_methods_stay_on_v1() {
-        let network = Network::WorldChain;
-        for method in [
-            RpcMethod::SponsorUserOperation,
-            RpcMethod::SendUserOperation,
-            RpcMethod::WaGetUserOperationReceipt,
-            RpcMethod::RelaySafeTransaction,
-            RpcMethod::SupportedEntryPoints,
-        ] {
-            let url = RpcClient::rpc_endpoint(network, &method);
-            assert!(
-                url.starts_with("/v1/"),
-                "{method:?} should route to /v1/, got {url}"
-            );
-        }
-    }
-
-    #[test]
-    fn test_rpc_endpoint_includes_network_name() {
-        let url = RpcClient::rpc_endpoint(
-            Network::WorldChain,
-            &RpcMethod::SendUserOperationV2,
-        );
-        assert_eq!(url, "/v2/rpc/worldchain");
-
-        let url =
-            RpcClient::rpc_endpoint(Network::WorldChain, &RpcMethod::SendUserOperation);
-        assert_eq!(url, "/v1/rpc/worldchain");
-    }
-
-    #[test]
-    fn test_pm_sponsor_response_parsing() {
-        // Bundler-sponsored shape — paymaster fields are absent from the
-        // response body entirely (not present-with-zero). Modelled as
-        // Option<T>, so all four deserialize to None.
-        let no_paymaster = json!({
-            "callGasLimit": "0x0",
-            "verificationGasLimit": "0x0",
-            "preVerificationGas": "0x0",
-            "maxFeePerGas": "0x0",
-            "maxPriorityFeePerGas": "0x0",
-        });
-        let r: PmSponsorUserOperationResponse =
-            serde_json::from_value(no_paymaster).unwrap();
-        assert_eq!(r.call_gas_limit, U128::ZERO);
-        assert_eq!(r.verification_gas_limit, U128::ZERO);
-        assert_eq!(r.pre_verification_gas, U256::ZERO);
-        assert_eq!(r.max_fee_per_gas, U128::ZERO);
-        assert_eq!(r.max_priority_fee_per_gas, U128::ZERO);
-        assert!(r.paymaster.is_none());
-        assert!(r.paymaster_verification_gas_limit.is_none());
-        assert!(r.paymaster_post_op_gas_limit.is_none());
-        assert!(r.paymaster_data.is_none());
-
-        // Self-sponsored shape — all four paymaster fields present with real
-        // values from Pimlico's ERC-20 paymaster.
-        let with_paymaster = json!({
-            "callGasLimit": "0x212df",
-            "verificationGasLimit": "0x501ab",
-            "preVerificationGas": "0x350f7",
-            "maxFeePerGas": "0x7A5CF70D5",
-            "maxPriorityFeePerGas": "0x3B9ACA00",
-            "paymaster": "0x0000000000000039cd5e8aE05257CE51C473ddd1",
-            "paymasterVerificationGasLimit": "0x6dae",
-            "paymasterPostOpGasLimit": "0x706e",
-            "paymasterData": "0x01000066d1a1a4",
-        });
-        let r: PmSponsorUserOperationResponse =
-            serde_json::from_value(with_paymaster).unwrap();
-        assert_eq!(
-            r.paymaster,
-            Some(address!("0000000000000039cd5e8aE05257CE51C473ddd1"))
-        );
-        assert_eq!(
-            r.paymaster_verification_gas_limit,
-            Some(U128::from(0x6dae_u32))
-        );
-        assert_eq!(r.paymaster_post_op_gas_limit, Some(U128::from(0x706e_u32)));
-        assert!(r.paymaster_data.is_some());
-    }
-
-    #[test]
-    fn test_sponsorship_context_serialization() {
-        // Protocol-sponsored: empty object is sent as the third param.
-        assert_eq!(SponsorshipContext::Protocol.to_json_value(), json!({}));
-
-        // Self-sponsored: { "token": "0x..." } with a lowercase 0x-prefixed
-        // hex address, matching the entry_point formatting convention.
-        let token = address!("2cfc85d8e48f8eab294be644d9e25c3030863003");
-        assert_eq!(
-            SponsorshipContext::SelfSponsoredToken(token).to_json_value(),
-            json!({ "token": "0x2cfc85d8e48f8eab294be644d9e25c3030863003" })
-        );
-    }
 }
